requirements.txt
faiss-cpu==1.10.0
google-generativeai==0.8.4
langchain-community==0.3.18
langchain-huggingface==0.1.2
python-dotenv==1.0.1
pypdf==5.3.1
streamlit==1.42.2